requirements.txt
asgiref==3.11.0
Django==5.2.8
et_xmlfile==2.0.0
numpy
openpyxl==3.1.5
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
xlrd==2.0.2
django-debug-toolbar
